requirements.txt
pandas
keras
tensorflow-cpu
matplotlib
numpy
jupyter
notebook
tqdm
kaggle
flask
opencv-python
scipy
mtcnn
fer
